requirements.txt
# Core
telethon==1.34.0
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.25
alembic==1.13.1
asyncpg==0.29.0
psycopg2-binary==2.9.9

# Redis
redis==5.0.1
aioredis==2.0.1

# LLM
openai==1.12.0

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1
httpx==0.26.0
tenacity==8.2.3
structlog==24.1.0

# Security
cryptography==42.0.2

# Testing
pytest==8.0.0
pytest-asyncio==0.23.4
pytest-cov==4.1.0

# Development
black==24.1.1
isort==5.13.2
mypy==1.8.0
